+++ src/Symfony/Bridge/Doctrine/Validator/Constraints/UniqueEntity.php
@@ -16,9 +16,6 @@ use Symfony\Component\Validator\Constraint;
 /**
  * Constraint for the Unique Entity validator.
  *
- * @Annotation
- * @Target({"CLASS", "ANNOTATION"})
- *
  * @author Benjamin Eberlei <kontakt@beberlei.de>
  */
 #[\Attribute(\Attribute::TARGET_CLASS | \Attribute::IS_REPEATABLE)]