+++ src/Symfony/Component/Validator/Constraints/Issn.php
@@ -14,9 +14,6 @@ namespace Symfony\Component\Validator\Constraints;
 use Symfony\Component\Validator\Constraint;
 
 /**
- * @Annotation
- * @Target({"PROPERTY", "METHOD", "ANNOTATION"})
- *
  * @author Antonio J. García Lagar <aj@garcialagar.es>
  * @author Bernhard Schussek <bschussek@gmail.com>
  */