+++ src/Symfony/Component/Validator/Mapping/Loader/AnnotationLoader.php
@@ -11,7 +11,6 @@
 
 namespace Symfony\Component\Validator\Mapping\Loader;
 
-use Doctrine\Common\Annotations\Reader;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Constraints\Callback;
 use Symfony\Component\Validator\Constraints\GroupSequence;
@@ -20,29 +19,13 @@ use Symfony\Component\Validator\Exception\MappingException;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
 
 /**
- * Loads validation metadata using a Doctrine annotation {@link Reader} or using PHP 8 attributes.
+ * Loads validation metadata using PHP attributes.
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  * @author Alexander M. Turek <me@derrabus.de>
  */
 class AnnotationLoader implements LoaderInterface
 {
-    /**
-     * @deprecated since Symfony 6.4, this property will be removed in 7.0
-     *
-     * @var Reader|null
-     */
-    protected $reader;
-
-    public function __construct(Reader $reader = null)
-    {
-        if ($reader) {
-            trigger_deprecation('symfony/validator', '6.4', 'Passing a "%s" instance as argument 1 to "%s()" is deprecated, pass null or omit the parameter instead.', get_debug_type($reader), __METHOD__);
-        }
-
-        $this->reader = $reader;
-    }
-
     public function loadClassMetadata(ClassMetadata $metadata): bool
     {
         $reflClass = $metadata->getReflectionClass();
@@ -98,49 +81,14 @@ class AnnotationLoader implements LoaderInterface
 
     private function getAnnotations(\ReflectionMethod|\ReflectionClass|\ReflectionProperty $reflection): iterable
     {
-        $dedup = [];
-
         foreach ($reflection->getAttributes(GroupSequence::class) as $attribute) {
-            $dedup[] = $attribute->newInstance();
             yield $attribute->newInstance();
         }
         foreach ($reflection->getAttributes(GroupSequenceProvider::class) as $attribute) {
-            $dedup[] = $attribute->newInstance();
             yield $attribute->newInstance();
         }
         foreach ($reflection->getAttributes(Constraint::class, \ReflectionAttribute::IS_INSTANCEOF) as $attribute) {
-            $dedup[] = $attribute->newInstance();
             yield $attribute->newInstance();
-        }
-        if (!$this->reader) {
-            return;
-        }
-
-        $annotations = [];
-
-        if ($reflection instanceof \ReflectionClass && $annotations = $this->reader->getClassAnnotations($reflection)) {
-            trigger_deprecation('symfony/validator', '6.4', 'Class "%s" uses Doctrine Annotations to configure validation constraints, which is deprecated. Use PHP attributes instead.', $reflection->getName());
-        }
-        if ($reflection instanceof \ReflectionMethod && $annotations = $this->reader->getMethodAnnotations($reflection)) {
-            trigger_deprecation('symfony/validator', '6.4', 'Method "%s::%s()" uses Doctrine Annotations to configure validation constraints, which is deprecated. Use PHP attributes instead.', $reflection->getDeclaringClass()->getName(), $reflection->getName());
-        }
-        if ($reflection instanceof \ReflectionProperty && $annotations = $this->reader->getPropertyAnnotations($reflection)) {
-            trigger_deprecation('symfony/validator', '6.4', 'Property "%s::$%s" uses Doctrine Annotations to configure validation constraints, which is deprecated. Use PHP attributes instead.', $reflection->getDeclaringClass()->getName(), $reflection->getName());
-        }
-
-        foreach ($dedup as $annotation) {
-            if ($annotation instanceof Constraint) {
-                $annotation->groups; // trigger initialization of the "groups" property
-            }
-        }
-
-        foreach ($annotations as $annotation) {
-            if ($annotation instanceof Constraint) {
-                $annotation->groups; // trigger initialization of the "groups" property
-            }
-            if (!\in_array($annotation, $dedup, false)) {
-                yield $annotation;
-            }
         }
     }
 }