+++ src/Symfony/Component/Validator/Tests/Constraints/WhenTest.php
@@ -11,9 +11,7 @@
 
 namespace Symfony\Component\Validator\Tests\Constraints;
 
-use Doctrine\Common\Annotations\AnnotationReader;
 use PHPUnit\Framework\TestCase;
-use Symfony\Bridge\PhpUnit\ExpectDeprecationTrait;
 use Symfony\Component\Validator\Constraints\Callback;
 use Symfony\Component\Validator\Constraints\NotBlank;
 use Symfony\Component\Validator\Constraints\NotNull;
@@ -26,8 +24,6 @@ use Symfony\Component\Validator\Tests\Constraints\Fixtures\WhenTestWithAttribute
 
 final class WhenTest extends TestCase
 {
-    use ExpectDeprecationTrait;
-
     public function testMissingOptionsExceptionIsThrown()
     {
         $this->expectException(MissingOptionsException::class);
@@ -43,89 +39,6 @@ final class WhenTest extends TestCase
         new When('true', [
             'foo',
         ]);
-    }
-
-    /**
-     * @group legacy
-     */
-    public function testAnnotations()
-    {
-        $this->expectDeprecation('Since symfony/validator 6.4: Passing a "Doctrine\Common\Annotations\AnnotationReader" instance as argument 1 to "Symfony\Component\Validator\Mapping\Loader\AnnotationLoader::__construct()" is deprecated, pass null or omit the parameter instead.');
-
-        $loader = new AnnotationLoader(new AnnotationReader());
-        $metadata = new ClassMetadata(WhenTestWithAnnotations::class);
-
-        $this->expectDeprecation('Since symfony/validator 6.4: Class "Symfony\Component\Validator\Tests\Constraints\WhenTestWithAnnotations" uses Doctrine Annotations to configure validation constraints, which is deprecated. Use PHP attributes instead.');
-        $this->expectDeprecation('Since symfony/validator 6.4: Property "Symfony\Component\Validator\Tests\Constraints\WhenTestWithAnnotations::$foo" uses Doctrine Annotations to configure validation constraints, which is deprecated. Use PHP attributes instead.');
-        $this->expectDeprecation('Since symfony/validator 6.4: Property "Symfony\Component\Validator\Tests\Constraints\WhenTestWithAnnotations::$bar" uses Doctrine Annotations to configure validation constraints, which is deprecated. Use PHP attributes instead.');
-        $this->expectDeprecation('Since symfony/validator 6.4: Property "Symfony\Component\Validator\Tests\Constraints\WhenTestWithAnnotations::$qux" uses Doctrine Annotations to configure validation constraints, which is deprecated. Use PHP attributes instead.');
-        $this->expectDeprecation('Since symfony/validator 6.4: Method "Symfony\Component\Validator\Tests\Constraints\WhenTestWithAnnotations::getBaz()" uses Doctrine Annotations to configure validation constraints, which is deprecated. Use PHP attributes instead.');
-
-        self::assertTrue($loader->loadClassMetadata($metadata));
-
-        [$classConstraint] = $metadata->getConstraints();
-
-        self::assertInstanceOf(When::class, $classConstraint);
-        self::assertSame('true', $classConstraint->expression);
-        self::assertEquals([
-            new Callback([
-                'callback' => 'callback',
-                'groups' => ['Default', 'WhenTestWithAnnotations'],
-            ]),
-        ], $classConstraint->constraints);
-
-        [$fooConstraint] = $metadata->properties['foo']->getConstraints();
-
-        self::assertInstanceOf(When::class, $fooConstraint);
-        self::assertSame('true', $fooConstraint->expression);
-        self::assertEquals([
-            new NotNull([
-                'groups' => ['Default', 'WhenTestWithAnnotations'],
-            ]),
-            new NotBlank([
-                'groups' => ['Default', 'WhenTestWithAnnotations'],
-            ]),
-        ], $fooConstraint->constraints);
-        self::assertSame(['Default', 'WhenTestWithAnnotations'], $fooConstraint->groups);
-
-        [$barConstraint] = $metadata->properties['bar']->getConstraints();
-
-        self::assertInstanceOf(When::class, $fooConstraint);
-        self::assertSame('false', $barConstraint->expression);
-        self::assertEquals([
-            new NotNull([
-                'groups' => ['foo'],
-            ]),
-            new NotBlank([
-                'groups' => ['foo'],
-            ]),
-        ], $barConstraint->constraints);
-        self::assertSame(['foo'], $barConstraint->groups);
-
-        [$quxConstraint] = $metadata->properties['qux']->getConstraints();
-
-        self::assertInstanceOf(When::class, $quxConstraint);
-        self::assertSame('true', $quxConstraint->expression);
-        self::assertEquals([
-            new NotNull([
-                'groups' => ['foo'],
-            ]),
-        ], $quxConstraint->constraints);
-        self::assertSame(['foo'], $quxConstraint->groups);
-
-        [$bazConstraint] = $metadata->getters['baz']->getConstraints();
-
-        self::assertInstanceOf(When::class, $bazConstraint);
-        self::assertSame('true', $bazConstraint->expression);
-        self::assertEquals([
-            new NotNull([
-                'groups' => ['Default', 'WhenTestWithAnnotations'],
-            ]),
-            new NotBlank([
-                'groups' => ['Default', 'WhenTestWithAnnotations'],
-            ]),
-        ], $bazConstraint->constraints);
-        self::assertSame(['Default', 'WhenTestWithAnnotations'], $bazConstraint->groups);
     }
 
     public function testAttributes()
@@ -198,38 +111,5 @@ final class WhenTest extends TestCase
             ]),
         ], $bazConstraint->constraints);
         self::assertSame(['Default', 'WhenTestWithAttributes'], $bazConstraint->groups);
-    }
-}
-
-/**
- * @When(expression="true", constraints={@Callback("callback")})
- */
-class WhenTestWithAnnotations
-{
-    /**
-     * @When(expression="true", constraints={@NotNull, @NotBlank})
-     */
-    private $foo;
-
-    /**
-     * @When(expression="false", constraints={@NotNull, @NotBlank}, groups={"foo"})
-     */
-    private $bar;
-
-    /**
-     * @When(expression="true", constraints=@NotNull, groups={"foo"})
-     */
-    private $qux;
-
-    /**
-     * @When(expression="true", constraints={@NotNull, @NotBlank})
-     */
-    public function getBaz()
-    {
-        return null;
-    }
-
-    public function callback()
-    {
     }
 }