+++ src/Symfony/Component/Validator/ValidatorBuilder.php
@@ -11,13 +11,8 @@
 
 namespace Symfony\Component\Validator;
 
-use Doctrine\Common\Annotations\AnnotationReader;
-use Doctrine\Common\Annotations\PsrCachedReader;
-use Doctrine\Common\Annotations\Reader;
 use Psr\Cache\CacheItemPoolInterface;
-use Symfony\Component\Cache\Adapter\ArrayAdapter;
 use Symfony\Component\Validator\Context\ExecutionContextFactory;
-use Symfony\Component\Validator\Exception\LogicException;
 use Symfony\Component\Validator\Exception\ValidatorException;
 use Symfony\Component\Validator\Mapping\Factory\LazyLoadingMetadataFactory;
 use Symfony\Component\Validator\Mapping\Factory\MetadataFactoryInterface;
@@ -48,7 +43,6 @@ class ValidatorBuilder
     private array $xmlMappings = [];
     private array $yamlMappings = [];
     private array $methodMappings = [];
-    private ?Reader $annotationReader = null;
     private bool $enableAnnotationMapping = false;
     private ?MetadataFactoryInterface $metadataFactory = null;
     private ConstraintValidatorFactoryInterface $validatorFactory;
@@ -210,35 +204,6 @@ class ValidatorBuilder
     public function disableAnnotationMapping(): static
     {
         $this->enableAnnotationMapping = false;
-        $this->annotationReader = null;
-
-        return $this;
-    }
-
-    /**
-     * @deprecated since Symfony 6.4 without replacement
-     *
-     * @return $this
-     */
-    public function setDoctrineAnnotationReader(?Reader $reader): static
-    {
-        trigger_deprecation('symfony/validator', '6.4', 'Method "%s()" is deprecated without replacement.', __METHOD__);
-
-        $this->annotationReader = $reader;
-
-        return $this;
-    }
-
-    /**
-     * @deprecated since Symfony 6.4 without replacement
-     *
-     * @return $this
-     */
-    public function addDefaultDoctrineAnnotationReader(): static
-    {
-        trigger_deprecation('symfony/validator', '6.4', 'Method "%s()" is deprecated without replacement.', __METHOD__);
-
-        $this->annotationReader = $this->createAnnotationReader();
 
         return $this;
     }
@@ -345,7 +310,7 @@ class ValidatorBuilder
         }
 
         if ($this->enableAnnotationMapping) {
-            $loaders[] = new AnnotationLoader($this->annotationReader);
+            $loaders[] = new AnnotationLoader();
         }
 
         return array_merge($loaders, $this->loaders);
@@ -388,18 +353,5 @@ class ValidatorBuilder
         $contextFactory = new ExecutionContextFactory($translator, $this->translationDomain);
 
         return new RecursiveValidator($contextFactory, $metadataFactory, $validatorFactory, $this->initializers);
-    }
-
-    private function createAnnotationReader(): Reader
-    {
-        if (!class_exists(AnnotationReader::class)) {
-            throw new LogicException('Enabling annotation based constraint mapping requires the packages doctrine/annotations and symfony/cache to be installed.');
-        }
-
-        if (class_exists(ArrayAdapter::class)) {
-            return new PsrCachedReader(new AnnotationReader(), new ArrayAdapter());
-        }
-
-        throw new LogicException('Enabling annotation based constraint mapping requires the packages doctrine/annotations and symfony/cache to be installed.');
     }
 }